requirements.txt
redis==2.10.6
pymongo==3.6.0
fabric3
requests==2.14.2
